Refactor and enhance app architecture
This commit significantly overhauls the application's architecture by introducing MediatR and AutoMapper to streamline object handling and reduce dependencies. We've refactored all command and query handlers to implement MediatR's IRequestHandler, ensuring a unified approach to request handling. Additionally, we've introduced DTOs for efficient data transfer and abstract records for commands and queries to leverage immutability and readability benefits of C# 9.

Validator classes have been added as placeholders for future input validation using FluentValidation, indicating a move towards robust data integrity checks. Error handling has been improved in handlers to catch and respond to not found scenarios more effectively. The repository pattern is now consistently applied across the application, enhancing maintainability and testability of data access layers.

Project structure has been reorganized for better file discoverability and to reduce clutter by removing unused folders. These changes collectively modernize the application, setting a solid foundation for future development and feature enhancements.

+++ src/MSL.Application/Features/Order/Commands/DeleteOrderCommand/DeleteOrderCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using MLS.Application.Contracts.Persistence.IRepositories;
+using MLS.Application.Exceptions;
 
 namespace MLS.Application.Features.Order.Commands.DeleteOrderCommand
 {
@@ -15,6 +16,10 @@ namespace MLS.Application.Features.Order.Commands.DeleteOrderCommand
         public async Task<Unit> Handle(DeleteOrderCommand request, CancellationToken cancellationToken)
         {
             var orderToDelete = await _orderRepository.GetByIdAsync(request.Id);
+
+            if (orderToDelete is null)
+                throw new NotFoundException(nameof(Domain.Entities.Order), request.Id);
+
             await _orderRepository.DeleteAsync(orderToDelete);
 
             return Unit.Value;
